Split targeting crate into library & binary
